scripts/requirements.txt
pyserial
argparse
hexdump
rich